% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_yo.R
\name{gate_fcs}
\alias{gate_fcs}
\title{Gate an FCS file}
\usage{
gate_fcs(
  fcs_data,
  name = "Flow Yo plot",
  x = "FSC-A",
  y = "SSC-A",
  filter_gate = data.frame(x = c(30000, 30000, 180000, 262143, 262143, 30000), y = c(0,
    40000, 262143, 262143, 0, 0)),
  gate_type = "polygon",
  gate_name = "gate",
  filter_type = "inside",
  log_x = FALSE,
  log_y = FALSE,
  max_x = 262143,
  max_y = 262143,
  plot_type = "hexagon"
)
}
\arguments{
\item{dataframe}{from read_fcs()}

\item{gate}{which can be a dataframe, a list or a number}

\item{logical}{if y axis should be logarithmic}

\item{numeric}{for the max y value of}

\item{string}{for the plot type}
}
\value{
A list containing the original dataframe with information about the
gate, a gated dataframe and a plot of the gating
}
\description{
This function takes and FCS dataframe from read_fcs() and a gate and returns
the original dataframe with information about the gate, a gated dataframe and
a plot of the gating.
}
