% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_yo.R
\name{read_fcs}
\alias{read_fcs}
\title{Load an FCS file}
\usage{
read_fcs(fcs_path)
}
\arguments{
\item{infile}{Path to the input FCS file}
}
\value{
A dataframe of the FCS infile
}
\description{
This function loads an FCS file as dataframe and returns it.
}
